requirements.txt
bidict            0.23.1
blinker           1.9.0
certifi           2026.2.25
click             8.1.7
cloudinary        1.44.1
colorama          0.4.6
dnspython         2.8.0
eventlet          0.40.4
Flask             3.1.3
flask-marshmallow 1.2.1
Flask-SocketIO    5.6.1
greenlet          3.3.2
h11               0.16.0
itsdangerous      2.2.0
Jinja2            3.1.3
jsonify           0.5
MarkupSafe        2.1.5
marshmallow       3.21.2
packaging         24.0
pip               21.3.1
pyodbc            5.1.0
python-engineio   4.13.1
python-socketio   5.16.1
setuptools        60.2.0
simple-websocket  1.1.0
six               1.17.0
urllib3           2.6.3
Werkzeug          3.1.6
wheel             0.37.1
wsproto           1.3.2
psycopg2-binary
python-dotenv